requirements.txt
pandas
plotly.express
matplotlib
streamlit-aggrid
streamlit-card
streamlit-extras